% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/idb.R
\name{idb1}
\alias{idb1}
\title{Retrieve data from the single-year-of-age IDB dataset}
\usage{
idb1(country, year, variables = "all", start_age = NULL, end_age = NULL,
  sex = NULL, api_key = NULL)
}
\arguments{
\item{country}{The two-character country FIPS code}

\item{year}{The year for which you'd like to retrieve data}

\item{variables}{A vector of variables.  If left blank, will return age, area in square kilometers, the name of the country, and the population size of the age group.}

\item{start_age}{(optional) The first age for which you'd like to retrieve data.}

\item{end_age}{(optional) The second age group for which you'd like to retrieve data.}

\item{sex}{(optional) One of 'both', 'male', or 'female'.}
}
\value{
A data frame with the requested data.
}
\description{
Retrieve data from the single-year-of-age IDB dataset
}

