% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/idb.R
\name{idb5}
\alias{idb5}
\title{Retrieve data from the five-year-age-group IDB dataset}
\usage{
idb5(country, year, variables = NULL, concept = NULL,
  country_name = FALSE, api_key = NULL)
}
\arguments{
\item{country}{A two-character FIPS code, or a vector of FIPS codes, of the countries for which you'd like to retrieve data.}

\item{year}{A year, or a vector of years, for which you'd like to retrieve data.}

\item{variables}{A vector of variables.  Use `idb_variables()` for a full list.}

\item{concept}{A concept for which you'd like to retrieve data.  Use `idb_concepts()` for a list of options.}

\item{country_name}{If TRUE, returns a column with the long country name along with the FIPS code.}
}
\value{
A data frame with the requested data.
}
\description{
Retrieve data from the five-year-age-group IDB dataset
}

