% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idb-deprecated.R
\name{idb_api_key}
\alias{idb_api_key}
\title{Set the Census API key}
\usage{
idb_api_key(api_key)
}
\arguments{
\item{api_key}{The idbr user's Census API key.  Can be obtained from \url{https://api.census.gov/data/key_signup.html}.}
}
\description{
Use to set the Census API key in an idbr session so that the key does not have to be passed to each
\code{idb1} or \code{idb5} function call.
}
\examples{
\dontrun{

idb_api_key('Your API key goes here')

}
}
