% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{idb_concepts}
\alias{idb_concepts}
\title{Print the available concepts to pass to the `idb5()` function.}
\usage{
idb_concepts()
}
\description{
Print the available concepts to pass to the `idb5()` function.
}
