% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{idb_variables}
\alias{idb_variables}
\title{Access IDB variables through api}
\usage{
idb_variables()
}
\value{
A tibble of 134 rows and 4 columns
}
\description{
Access IDB variables through api
}
\examples{
idb_variables()
}
