% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idbr-package.R
\docType{package}
\name{idbr-package}
\alias{idbr-package}
\title{idbr: R Interface to the US Census Bureau International Data Base API}
\description{
Use R to make requests to the US Census Bureau's International Data Base API. Results are returned as R data frames. For more information about the IDB API, visit \url{https://www.census.gov/data/developers/data-sets/international-database.html}.
}
\author{
\strong{Maintainer}: Kyle Walker \email{kyle.walker@tcu.edu}

}
\keyword{internal}
