% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/idbr_package.R
\docType{package}
\name{idbr}
\alias{idbr}
\alias{idbr-package}
\title{Access the Census International Data Base (IDB) from R}
\description{
This R package grants users access to the US Census Bureau's International Data Base API, and returns queries
as R data frames.
}
\note{
This product uses the Census Bureau Data API but is not endorsed or certified by the Census Bureau.
}
\author{
Kyle Walker
}
\seealso{
Census API terms of service: \url{http://www.census.gov/data/developers/about/terms-of-service.html}
}

