% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_idb.R
\name{get_idb}
\alias{get_idb}
\title{Get Data from the US Census Bureau's International Data Base API}
\usage{
get_idb(
  country,
  year,
  variables = NULL,
  age = NULL,
  sex = NULL,
  geometry = FALSE,
  resolution = c("low", "high"),
  api_key = NULL
)
}
\arguments{
\item{country}{A country name or vector of country names.}

\item{year}{A year, or vector of years}

\item{variables}{The variables you'd like to request.  If filtering by age or sex, should be NULL.}

\item{age}{A vector of ages}

\item{sex}{Either NULL (both sexes, the default), "male", or "female"}

\item{geometry}{Not yet implemented}

\item{resolution}{"high" or "low"}

\item{api_key}{Your key}
}
\value{
Data from the IDB!
}
\description{
Get Data from the US Census Bureau's International Data Base API
}
