% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{idb_variables}
\alias{idb_variables}
\title{Print the available variables to pass to the `idb5()` function.}
\usage{
idb_variables()
}
\description{
The first column, "Name", details the variable names that can be passed the function.  The second column, "Label", describes the content of the variables.
}
