% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/idbr_package.R
\docType{package}
\name{idbr}
\alias{idbr}
\alias{idbr-package}
\title{Access the Census International Data Base (IDB) from R}
\description{
This R package grants users access to the US Census Bureau's International Data Base (IDB) API, and returns queries as R data frames.  The IDB includes historical demographic data, current population estimates, and demographic projections to 2050 for countries of population 5,000 or greater that are recognized by the US Department of State.  Demographic indicators in the IDB include mid-year population; population counts by sex and age; and fertility, mortality, and migration variables such as net migration, infant mortality rates, and total fertility rates.  Future projections of these indicators are estimated using the cohort-component method.  For details on the US Census Bureau's methodology for producing population estimates, please visit \url{http://www.census.gov/population/international/data/idb/estandproj.pdf}.
}
\note{
This product uses the Census Bureau Data API but is not endorsed or certified by the Census Bureau.
}
\author{
Kyle Walker
}
\seealso{
Census API terms of service: \url{http://www.census.gov/data/developers/about/terms-of-service.html}

US Census Bureau IDB API home page: \url{http://www.census.gov/data/developers/data-sets/international-database.html}
}

