% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IDBr_package.R
\docType{package}
\name{IDBr}
\alias{IDBr}
\alias{IDBr-package}
\title{Access the Census International Data Base (IDB) from R}
\description{
Access the Census International Data Base (IDB) from R
}
\author{
Kyle Walker
}

