% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/idbr_package.R
\docType{package}
\name{idbr}
\alias{idbr}
\alias{idbr-package}
\title{Access the Census International Data Base (IDB) from R}
\description{
Access the Census International Data Base (IDB) from R
}
\author{
Kyle Walker
}

